refactor: abstract services into separate module outside of manager
